🎨 nest app 셋팅 한 곳에서 하게 설정 및 컨벤션 맞지않은 파일명 변경
